# backend/docs/swagger.yaml
basePath: /
definitions:
  api.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  api.DoorStateResponse:
    properties:
      expected_angle:
        example: 0
        type: integer
      state:
        example: LOCKED
        type: string
    type: object
  api.SignInRequest:
    properties:
      name:
        type: string
      password:
        type: string
    required:
    - name
    - password
    type: object
  api.SignUpRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend API for the Smart Door Security system. Manages door access,
    face recognition, intrusion detection, and MQTT-based IoT communication.
  title: Smart Door Security API
  version: "1.0"
paths:
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticates by name + password and returns a JWT
      parameters:
      - description: Sign-in credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign in
      tags:
      - auth
  /auth/signout:
    post:
      description: Invalidates the client session. Because JWTs are stateless, the
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Sign out
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a new user account and returns a JWT
      parameters:
      - description: Sign-up credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /door/lock:
    post:
      description: Sends an MQTT LOCK command to the servo. Logs a MANUAL_LOCK event.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lock door
      tags:
      - door
  /door/state:
    get:
      description: Returns the servo angle and lock state last commanded by the backend
        (LOCKED / UNLOCKED).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DoorStateResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current door lock state
      tags:
      - door
  /door/unlock:
    post:
      description: Sends an MQTT UNLOCK command to the servo and auto-locks after
        5 s. Logs a MANUAL_UNLOCK event.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unlock door (manual)
      tags:
      - door
  /events:
    get:
      description: Returns the most recent events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List events
      tags:
      - events
  /events/{id}:
    get:
      description: Returns a single event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get event by ID
      tags:
      - events
  /family:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List family members
      tags:
      - family
    post:
      consumes:
      - application/json
      parameters:
      - description: name
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      summary: Create a family member (name only; enroll face separately)
      tags:
      - family
  /family/{id}:
    delete:
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a family member and remove their face enrollment
      tags:
      - family
    get:
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a family member by ID
      tags:
      - family
    put:
      consumes:
      - application/json
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Rename a family member
      tags:
      - family
  /family/{id}/enroll:
    delete:
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Remove the face enrollment for a member
      tags:
      - family
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      - description: Face photo
        in: formData
        name: photo
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Upload a photo and enroll the member's face in the recognition model
      tags:
      - family
  /health:
    get:
      description: Returns the health status of the API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
  /users:
    get:
      description: Returns a list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a new user with name, email, and password
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {your-jwt-token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"